refactor(Info): update imports to remove circular dependency
