+++ packages/core/src/components/Info/Info.tsx
@@ -1,6 +1,7 @@
 import React, { forwardRef, useCallback, useRef, useState } from "react";
 import { Info as InfoIcon } from "@vibe/icons";
-import { IconButton, Dialog } from "../../";
+import { IconButton } from "../IconButton";
+import { Dialog } from "../Dialog";
 import { InfoDialogContent } from "./components";
 import { type InfoProps } from "./Info.types";
 import { ComponentDefaultTestId, ComponentVibeId } from "../../tests/constants";
